refactor(app): extract the SQL editor wiring
The Rust-owned text buffer, the lexer feeding the span model, statement
folding and autocomplete move to `wire::editor`.

This module returns two handles rather than only installing callbacks:
`sync_editor` and `load_editor_text` are built here but needed almost
everywhere, so `main` takes them back and passes them on through
`AppFns`.
